+++ src/about.cpp
@@ -22,8 +22,8 @@ FILDEF void do_about ()
     float lw = get_viewport().w - (XPAD*2.0f);
     float lh = 18.0f;
 
-    do_label          (UI_ALIGN_LEFT,UI_ALIGN_CENTER, lw,lh, format_string("The End is Nigh Editor (v%d.%d.%d) %s Build", EDITOR_MAJOR,EDITOR_MINOR,EDITOR_PATCH, build).c_str());
-    do_label_hyperlink(UI_ALIGN_LEFT,UI_ALIGN_CENTER, lw,lh, "Copyright (C) 2019 ", "Joshua Robertson", "http://that-boy.com/");
+    do_label_hyperlink(UI_ALIGN_LEFT,UI_ALIGN_CENTER, lw,lh, format_string("The End is Nigh Editor (v%d.%d.%d) %s Build ", EDITOR_MAJOR,EDITOR_MINOR,EDITOR_PATCH, build).c_str(), "Changelog", "https://github.com/thatb0y/tein-editor/blob/master/CHANGES.md");
+    do_label          (UI_ALIGN_LEFT,UI_ALIGN_CENTER, lw,lh, "Copyright (C) 2019 Joshua Robertson");
     advance_panel_cursor(lh/2.0f);
     do_label_hyperlink(UI_ALIGN_LEFT,UI_ALIGN_CENTER, lw,lh, "Flame icon made by Those Icons from ", "Flaticon", "https://www.flaticon.com/");