+++ CHANGES.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## v1.0.1
+* Sub-windows no longer appear in taskbar or ALT+TAB list.
+* Changed the maximum level size to be 2840x2840.
+* Changelog accessible via the About window.
+* Can now run the game from the editor.
+* Fixed About window text being cut-off when using OpenDyslexic.
+* Fixed hotkey text being cut-off when using OpenDyslexic.
+
 ## v1.0.0
 * Initial release.